+++ src/main/java/com/dash/leap/domain/mission/entity/Mission.java
@@ -1,6 +1,6 @@
 package com.dash.leap.domain.mission.entity;
 
-import com.dash.leap.config.domain.BaseEntity;
+import com.dash.leap.global.domain.BaseEntity;
 import com.dash.leap.domain.mission.entity.enums.MissionType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
